Fix triangle winding of vertical redstone wire
There is Z fighting between blocks and attached vertical redstone wire.
This is because the Z coordinate is interpolated linearly in screen space,
not accounting for perspective calculation.
                                                       _
Blocks were drawn like |\  but the attached wire like | /.
                       |_\                            |/

By using the same triangle winding for blocks and wire, it seems like the
interpolation error of both is similar, so Z fighting is avoided.